This is a major commit. The code now compiles both with Visual Studio 2008 and gcc (albeit version 4.0.1 only) without any warnings, i.e. with /W4 /Wp64 for VS2008 and -Wall -Wextra for gcc.

+++ test/test_ami_pmerge.cpp
@@ -149,7 +149,7 @@ ami::err s_merge_manager::operate(int **in,
         *out = min_t;
         if (in[min_source] != NULL) {
             mheap->insert(in[min_source], min_source);
-            taken_index = min_source;
+            taken_index = (int)min_source;
             //taken_flags[min_source] = 1;
 #if DEBUG_ASSERTIONS
             input_count++;
